WIP #1 git reflow
Moved around project controller methods
Removed older project table settings
Added in new routes for push and pull (ajax only)
Changed side menu and permissions
Added way in project to get structured git export and import info and do git checks

next need to do js handlers for push and pull buttons and test them out
