+++ database/hexlet_migrations/20210708204104_add_flow_projects.php
@@ -33,18 +33,11 @@ class AddFlowProjects extends AbstractMigration
                 `parent_flow_project_id` INT NULL DEFAULT NULL, 
                 `created_at_ts` INT NULL DEFAULT NULL,
                 `is_public` TINYINT NOT NULL DEFAULT 0,
-                `export_repo_do_auto_push` TINYINT NOT NULL DEFAULT 0,
                 `flow_project_guid` BINARY(16) NOT NULL ,
                 `flow_project_type` VARCHAR(10) NULL DEFAULT NULL ,
                 `flow_project_special_flag` VARCHAR(10) DEFAULT NULL,
                 `flow_project_title` VARCHAR(40) NULL DEFAULT NULL ,
                 `flow_project_blurb` VARCHAR(120) NULL DEFAULT NULL ,
-                `export_repo_url` VARCHAR(200) NULL DEFAULT NULL ,
-                `export_repo_branch` VARCHAR(200) NULL DEFAULT NULL ,
-                `export_repo_key` TEXT NULL DEFAULT NULL ,
-                `import_repo_url` VARCHAR(200) NULL DEFAULT NULL ,
-                `import_repo_branch` VARCHAR(200) NULL DEFAULT NULL ,
-                `import_repo_key` TEXT NULL DEFAULT NULL ,
                 `flow_project_readme` MEDIUMTEXT NULL DEFAULT NULL ,
                 `flow_project_readme_bb_code` MEDIUMTEXT NULL DEFAULT NULL ,
                 PRIMARY KEY (`id`),
