+++ views/pages/project/edit_project.twig
@@ -3,11 +3,14 @@
 <div class="row">
     <div class="col-12 col-lg-auto ">
         {{ include('project/partials/edit-side-menu.twig') }}
+        {% include('project/partials/permissions_link.twig') %}
     </div><!-- /.col -->
 
-    <div class="col ">
-        {{ include('project/partials/project-form.twig') }}
+    <div class="col col-lg-9 ">
+
+        {% include('project/partials/project-form.twig') %}
         <div class="w-100 mt-2"></div>
+
         {% include 'tag/partials/applied-control.twig' with{
             'tags_of_applied': project.get_applied_tags(),
             'applied_target_guid':project.flow_project_guid,
