+++ views/pages/project/new_project.twig
@@ -11,7 +11,7 @@
 
 <div class="row">
 
-    <div class="col ">
+    <div class="col col-lg-9 ">
         {{ include('project/partials/project-form.twig') }}
     </div><!-- /.col -->
