+++ views/pages/project/partials/history/export-import.twig
@@ -8,6 +8,52 @@
 
         <div class="card-body flow-import-export-settings" >
             <div class="row">
+
+                <div class="col-3">
+                    {% if project.get_current_user_permissions().can_edit %}
+                        <div class="d-inline-block">
+                            <button type="button" class="btn btn-outline-primary py-0 flow-git-pull-now" >
+                                Pull Now
+                            </button>
+                        </div>
+                    {% endif %} {# end can-edit #}
+                </div> <!--.col  -->
+
+                <div class="col-3">
+                    {% if project.get_current_user_permissions().can_read %}
+                        <div class="" aria-current="page">
+                            <a href="{{ url_for('download_project_export',{
+                                user_name: project.get_admin_user().flow_user_name ,
+                                project_name: project.flow_project_title
+                            } )
+                            }}"
+                               class="btn btn-outline-info py-0"
+                            >
+                                Download as Zip
+                            </a>
+                        </div>
+                    {% endif %} {# end can-read #}
+                </div> <!-- .col -->
+
+
+
+                <div class="col-3">
+                    {% if project.get_current_user_permissions().can_edit %}
+                        <div class="d-inline-block">
+                            <button type="button" class="btn btn-outline-primary py-0 flow-git-push-now" >
+                                Push Now
+                            </button>
+                        </div>
+                    {% endif %} {# end can-edit #}
+                </div> <!--.col  -->
+
+            </div> <!-- .row -->
+        </div> <!-- .card-body -->
+
+
+
+        <div class="card-body flow-import-export-settings" >
+            <div class="row">
                 <div class="col-6">
                     <div class="flow-git-import-here"></div>
                 </div>
