--- views/pages/project/partials/import/import-from-git.twig
@@ -1,62 +1,0 @@
-<div class="container">
-    <div class="row">
-        <div class="col-12 col-sm-12 col-md-10 offset-md-1 col-lg-10 offset-lg-1 col-xl-10 offset-xl-1">
-            <div class=" bg-light mt-3 px-4">
-                <form method="post" action="{{-  url_for('project_import_from_git',{
-                    user_name: project.get_admin_user().flow_user_name ,
-                    project_name: project.flow_project_title
-                } ) -}}"
-                      class="row g-3"
-                >
-                    {{ form_token(  url_for('project_import_from_git',{
-                        user_name: project.get_admin_user().flow_user_name ,
-                        project_name: project.flow_project_title
-                    } )) }}
-
-                    <input type="hidden" name="flow_project_git_hash" value="{{ project.getFlowProjectFiles().get_head_commit_hash() }}">
-
-
-                    <div class="d-inline-block float-start">
-                        <button type="submit" class="btn btn-success" name="import-now" value="1">
-                            Pull Now (remote must have history of current)
-                        </button>
-                    </div>
-
-                    <div class="col-12">
-                        <label for="import_repo_url">
-                            Repository Url
-                            <small class="p-2">
-                                Github only for right now
-                            </small>
-                        </label>
-                        <input type="text" id="import_repo_url" name="import_repo_url"
-                               class="form-control" placeholder="import Repo Url"  autofocus
-                               value="{{ project.import_repo_url }}"
-                        >
-                    </div>
-
-
-                    <div class="col-12">
-                        <label for="import_repo_branch">Repository Branch</label>
-                        <input type="text" id="import_repo_branch" name="import_repo_branch"
-                               class="form-control" placeholder="import Repo Branch"  autofocus
-                               value="{{ project.import_repo_branch }}"
-                        >
-                    </div>
-
-
-
-                    <div class="col-12">
-                        <label for="import_repo_key">Private Repo Key (no passphrase)</label>
-                        <textarea type="text" id="import_repo_key" name="import_repo_key"
-                                  class="form-control" placeholder="Private Repo Key"  autofocus
-                                  rows="15"
-                        >{{ project.import_repo_key }}</textarea>
-                    </div>
-
-
-                </form>
-            </div>
-        </div>
-    </div>
-</div>
