--- views/pages/project/project_export.twig
@@ -1,37 +1,0 @@
-{{ include('project/partials/project-breadcrumbs.twig') }}
-
-<div class="float-end">
-    {% if project.get_current_user_permissions().can_edit %}
-        <div class="breadcrumb-item active" aria-current="page">
-            <a href="{{ url_for('download_project_export',{
-                user_name: project.get_admin_user().flow_user_name ,
-                project_name: project.flow_project_title
-            } )
-            }}"
-               class="btn btn-outline-info py-0"
-            >
-                Download as Zip
-            </a>
-        </div>
-    {% endif %}
-</div>
-
-
-
-
-<div class="row">
-    <div class="col-12 col-lg-auto ">
-        {{ include('project/partials/edit-side-menu.twig') }}
-    </div><!-- /.col -->
-
-    {% if project.get_current_user_permissions().can_edit %}
-        <div class="col ">
-            {{ include('project/partials/export/export-push.twig') }}
-        </div><!-- /.col -->
-    {% endif %}
-
-
-</div> <!-- /.row -->
-
-
-
