+++ views/pages/project/project_history.twig
@@ -7,7 +7,7 @@
         {{ include('project/partials/edit-side-menu.twig') }}
     </div><!-- /.col -->
 
-    <div class="col ">
+    <div class="col col-lg-9">
         {{ include('project/partials/history/export-import.twig') }}
         {{ include('project/partials/history/history.twig') }}
     </div><!-- /.col -->
