--- views/pages/project/project_import.twig
@@ -1,30 +1,0 @@
-{{ include('project/partials/project-breadcrumbs.twig') }}
-
-
-
-<div class="row">
-    <div class="col-12 col-lg-auto ">
-        {{ include('project/partials/edit-side-menu.twig') }}
-    </div><!-- /.col -->
-
-    <div class="col ">
-        <div class="row">
-            {% if project.get_current_user_permissions().can_admin %}
-                <div class="col-12 col-xl-8">
-                    {{ include('project/partials/import/import-from-git.twig') }}
-                </div><!-- /.col -->
-            {% endif %}
-
-
-{#            {% if project.get_current_user_permissions().can_admin %}#}
-{#                <div class="col-12 col-xl-4 ">#}
-{#                    {{ include('project/partials/import/import-from-file.twig') }}#}
-{#                </div><!-- /.col -->#}
-{#            {% endif %}#}
-        </div> <!-- /.row -->
-    </div><!-- /.col -->
-
-</div> <!-- /.row -->
-
-
-
