+++ views/pages/project/resources.twig
@@ -6,7 +6,7 @@
         {{ include('project/partials/edit-side-menu.twig') }}
     </div><!-- /.col -->
 
-    <div class="col ">
+    <div class="col col-lg-9">
 
         <div class="container p-0 m-1">
