+++ views/pages/project/single_project_home.twig
@@ -7,7 +7,7 @@
         {{ include('project/partials/edit-side-menu.twig') }}
     </div><!-- /.col -->
 
-    <div class="col ">
+    <div class="col col-lg-9 ">
         {{ include('project/partials/display-project.twig') }}
     </div><!-- /.col -->
